Show output port path on SVG connection labels (#2878)
The flowc SVG renderer was only showing the connection name on edge
labels, ignoring the output port. Extract connection_label() into
flowcore so both flowc and flowedit share the same logic that combines
port name and connection name (e.g. "feedback-step /i2").
